Click and Position/qidian/requirements.txt
Pillow
opencv-python
configargparse
pyyaml
